refactored header structure for single import statement into source code via #include lib.h
