+++ include/hashTable.h
@@ -1,3 +1,5 @@
+#pragma once
+#include "value.h"
 #include "autoGrad.h"
 
 /**
